рефакторинг карточки места
