+++ lib/features/common/widgets/for_card/card_content_type.dart
